Bodyprog: matches (#156)
* Match `func_80086470`

* Match `func_800865FC`

* Match `func_800877B8`

* Match `func_800881B8`

* Match `func_80087AF4`

+++ src/bodyprog/bodyprog_8004A87C.c
@@ -296,7 +296,7 @@ void func_80054928() // 0x80054928
 
 INCLUDE_ASM("asm/bodyprog/nonmatchings/bodyprog_8004A87C", func_800549A0);
 
-void func_80054A04(s8 arg0) // 0x80054A04
+void func_80054A04(s32 arg0) // 0x80054A04
 {
     #define ADDR (void*)0x801E3600
